requirements.txt
asn1crypto==0.24.0
asgiref
# attrs==17.4.0
attrs
Automat==0.6.0
backcall==0.2.0
blinker==1.4
celery==4.4.7
certifi==2018.1.18
chardet==3.0.4
click==6.7
# cloud-init==0.3.0
colorama==0.3.7
# command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
# cryptography==2.1.4
decorator==5.0.7
django==3.2.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
# distro-info==0.18
# distro-info
# ec2-hibinit-agent==1.0.0
encoding-tools==0.0.3
# hibagent==1.0.1
# hibagent
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keyring==10.6.0
keyrings.alt==3.0
# language-selector==0.1
# MarkupSafe==1.0
mysqlclient==2.0.3
MarkupSafe
# netifaces==0.10.4
netifaces
oauthlib==2.0.6
# PAM==0.4.2
PAM
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
postgres==3.0.0
postgres.client==0.4.3
prompt-toolkit==3.0.18
psycopg2
#psycopg2==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
django-rest-swagger==2.2.0
drf-yasg

pyDes==2.0.1
Pygments==2.8.1
# pygobject==3.26.1
PyJWT==1.5.3
pymongo==3.11.0
pyngrok==5.0.5
pyOpenSSL==17.5.0
pyserial==3.4
pytest==6.2.4
# python-apt==1.6.5
python-dateutil==2.8.1
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
# ssh-import-id==5.7
ssh-import-id
# systemd-python==234
tornado==6.1
traitlets==4.3.3
# Twisted==17.9.0
# ufw==0.36
# unattended-upgrades==0.1
urllib3==1.22
# virtualenv==15.1.0
wcwidth==0.2.5
XlsxWriter==1.3.3
zope.interface==5.4.0



pytz==2020.4
sqlparse==0.4.1
parsedatetime
psycopg2-binary
amqp==2.6.1
billiard==3.6.3.0
coreapi==2.3.3
#nltk==3.5
numpy==1.19.1
openpyxl==3.0.4
python-crontab==2.5.1
pandas==1.1.1
PyMySQL==1.0.2
xlrd==1.2.0
# Cython==0.29.21
# et-xmlfile==1.0.1
# flower==0.9.5
# future==0.18.2
# importlib-metadata==1.7.0
# jdcal==1.4.1
# joblib==0.16.0
# kombu==4.6.11
# # mysqlclient
# oschmod==0.3.7
# #pyfcm==1.5.1
# regex==2020.7.14
# simple-crypt==4.1.7
# tqdm==4.48.2
# vine==1.3.0
# zipp==3.1.0

